Add extra digicert intermediate cert for win7
timestamp.digicert.com recently started returning signatures chaining to
the "DigiCert Trusted Root G4" root.  Because that root is not included
on old Windows 7 versions, they can't verify the timestamp on our recent
installers.

To work around this for now, add an extra intermediate certificate to
the chain so we get back to the old "DigiCert Global Root CA" root.